+++ src/index.ts
@@ -1,3 +1,5 @@
 export * from './slack-events';
 export * from './slack-textract';
+export * from './static-website';
+export * from './state-machine-cr-provider';
 export * from './email-receiver';