+++ CHANGELOG.md
@@ -2,6 +2,52 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### 0.1.2 (2020-12-07)
+
+### 0.1.1 (2020-12-07)
+
+## 0.1.0 (2020-12-07)
+
+
+### ⚠ BREAKING CHANGES
+
+* **slack-events:** Event subscriptions of existing Slack apps must be reconnected to the new endpoint.
+* **slack-events**: `SlackEventsProps.restApiName` is now `SlackEventsProps.apiName`
+
+### Features
+
+* **slack-events:** use a HTTP API ([#23](https://github.com/jogold/cloudstructs/issues/23)) ([95e1ed3](https://github.com/jogold/cloudstructs/commit/95e1ed3008a70ec347f12bc2ad9eb103149647f1))
+
+### 0.0.12 (2020-12-07)
+
+### 0.0.11 (2020-12-01)
+
+### 0.0.10 (2020-11-30)
+
+
+### Features
+
+* **state-machine-cr-provider:** custom resources with state machines ([#22](https://github.com/jogold/cloudstructs/issues/22)) ([ba4a814](https://github.com/jogold/cloudstructs/commit/ba4a814687f86e582c6362babcbce09547521be7))
+
+### 0.0.9 (2020-11-30)
+
+### 0.0.8 (2020-11-29)
+
+
+### Features
+
+* **static-website:** static website with backend configuration ([#18](https://github.com/jogold/cloudstructs/issues/18)) ([2a7a429](https://github.com/jogold/cloudstructs/commit/2a7a42915316117ba4505e00b9b49ea39dc176e5))
+
+### 0.0.7 (2020-11-29)
+
+### 0.0.6 (2020-11-27)
+
+### 0.0.5 (2020-11-26)
+
+### 0.0.4 (2020-11-26)
+
+### 0.0.3 (2020-11-25)
+
 ### 0.0.2 (2020-11-25)